refactor(goes-dl): relocate geodesy and netcdf
