+++ src/goesdl/geodesy/helpers.py
@@ -22,7 +22,7 @@ from numpy import (
     where,
 )
 
-from ..array import ArrayFloat64
+from ..utils.array import ArrayFloat64
 
 
 def make_consistent(
