+++ src/goesdl/netcdf/__init__.py
