+++ src/goesdl/netcdf/annotations.py
