+++ src/goesdl/netcdf/class_help.py
