+++ src/goesdl/netcdf/fields.py
