+++ src/goesdl/netcdf/hinting.py
