+++ src/goesdl/netcdf/validation.py
