+++ src/goesdl/netcdf/view.py
