+++ src/goesdl/utils/array.py
